OSCAR-361
Added a readyCallbacks property to the oscar object.

Added an _isReady function to hte oscar object. This is an internal function that is not to be called directly.

Replaced the onready function with onReady. Takes a function as a parameter which will be added to
the readyCallbacks property. These callbacks will be called when the loading has finished.

Added ability to load external scripts.