SFV-1121
The KML poi layer will use the POI icons if the geometries from the feed are not points. Otherwise it will extract the style information from the kml feed itself.

Updated the parseText utility to return an empty string if the input value is null, this prevents undefined from being displayed in the popup.